requirements.txt
colorama==0.4.6
requests==2.31.0
python-nmap==0.7.1
dnspython==2.4.2
python-whois==0.9.0
python-dotenv==1.0.0